chore(refactor): cli -> cmd
