+++ cmd/launcher/icon.go
