+++ cmd/launcher/internal/release_manager_test.go
@@ -8,7 +8,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	launcher "github.com/mudler/LocalAI/cli/launcher/internal"
+	launcher "github.com/mudler/LocalAI/cmd/launcher/internal"
 )
 
 var _ = Describe("ReleaseManager", func() {
