+++ cmd/launcher/internal/systray_manager.go
